修复若干bug； 添加一键下载json接口；

+++ src/main/java/me/zhouzhuo810/zzapidoc/project/service/impl/InterfaceServiceImpl.java
@@ -7,19 +7,30 @@ import me.zhouzhuo810.zzapidoc.common.utils.DataUtils;
 import me.zhouzhuo810.zzapidoc.common.utils.MapUtils;
 import me.zhouzhuo810.zzapidoc.project.dao.InterfaceDao;
 import me.zhouzhuo810.zzapidoc.project.entity.InterfaceEntity;
+import me.zhouzhuo810.zzapidoc.project.entity.ProjectEntity;
 import me.zhouzhuo810.zzapidoc.project.entity.RequestArgEntity;
 import me.zhouzhuo810.zzapidoc.project.entity.ResponseArgEntity;
 import me.zhouzhuo810.zzapidoc.project.service.InterfaceService;
+import me.zhouzhuo810.zzapidoc.project.service.ProjectService;
 import me.zhouzhuo810.zzapidoc.project.service.RequestArgService;
 import me.zhouzhuo810.zzapidoc.project.service.ResponseArgService;
 import me.zhouzhuo810.zzapidoc.project.utils.InterfaceUtils;
+import me.zhouzhuo810.zzapidoc.project.utils.ResponseArgUtils;
 import me.zhouzhuo810.zzapidoc.user.entity.UserEntity;
 import me.zhouzhuo810.zzapidoc.user.service.UserService;
+import org.apache.commons.io.FileUtils;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.json.JSONStringer;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.io.File;
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -39,6 +50,9 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
 
     @Resource(name = "responseArgServiceImpl")
     ResponseArgService mResponseArgService;
+
+    @Resource(name = "projectServiceImpl")
+    ProjectService mProjectService;
 
     @Override
     @Resource(name = "interfaceDaoImpl")
@@ -69,7 +83,7 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
             getBaseDao().save(entity);
             MapUtils map = new MapUtils();
             map.put("id", entity.getId());
-            if (requestArgs != null) {
+            if (requestArgs != null&&requestArgs.length()>0) {
                 JSONArray array = new JSONArray(requestArgs);
                 for (int i = 0; i < array.length(); i++) {
                     JSONObject jsonObject = array.getJSONObject(i);
@@ -88,11 +102,11 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
                     mRequestArgService.save(args);
                 }
             }
-            if (responseArgs != null) {
+            if (responseArgs != null&&responseArgs.length()>0) {
                 JSONArray array = new JSONArray(responseArgs);
                 for (int i = 0; i < array.length(); i++) {
                     JSONObject jsonObject = array.getJSONObject(i);
-                    saveResponseArg(jsonObject, projectId, user.getId(), user.getName(), "0");
+                    saveResponseArg(jsonObject, projectId, entity.getId(), user.getId(), user.getName(), "0");
                 }
             }
             return new BaseResult(1, "添加成功！", map.build());
@@ -102,7 +116,7 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
         }
     }
 
-    private void saveResponseArg(JSONObject jsonObject, String projectId, String userId, String userName, String pid) {
+    private void saveResponseArg(JSONObject jsonObject, String projectId, String interfaceId, String userId, String userName, String pid) {
         String argName = jsonObject.getString("name");
         int typeId = jsonObject.getInt("typeId");
         String argNote = jsonObject.getString("note");
@@ -112,6 +126,7 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
         args.setProjectId(projectId);
         args.setTypeId(typeId);
         args.setNote(argNote);
+        args.setInterfaceId(interfaceId);
         args.setCreateUserID(userId);
         args.setCreateUserName(userName);
         try {
@@ -126,13 +141,13 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
                     break;
                 case ResponseArgEntity.TYPE_OBJECT:
                     JSONObject child = jsonObject.getJSONObject("child");
-                    saveResponseArg(child, projectId, userId, userName, id);
+                    saveResponseArg(child, projectId, interfaceId, userId, userName, id);
                     break;
                 case ResponseArgEntity.TYPE_ARRAY_OBJECT:
                     JSONArray childs = jsonObject.getJSONArray("child");
                     for (int i = 0; i < childs.length(); i++) {
                         JSONObject jsonObject1 = childs.getJSONObject(i);
-                        saveResponseArg(jsonObject1, projectId, userId, userName, id);
+                        saveResponseArg(jsonObject1, projectId, interfaceId, userId, userName, id);
                     }
                     break;
                 case ResponseArgEntity.TYPE_ARRAY_STRING:
@@ -173,6 +188,8 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
     public BaseResult updateInterface(String interfaceId, String name, String path, String projectId, String groupId, String httpMethodId,
                                       String note, String userId, String requestArgs, String responseArgs) {
         UserEntity user = mUserService.get(userId);
+        if (user == null)
+            return new BaseResult(0, "用户不合法！");
         InterfaceEntity entity = getBaseDao().get(interfaceId);
         if (entity == null) {
             return new BaseResult(0, "该接口不存在或已被删除！");
@@ -183,11 +200,9 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
         entity.setGroupId(groupId);
         entity.setHttpMethodId(httpMethodId);
         entity.setNote(note);
-        if (user != null) {
-            entity.setModifyUserID(user.getId());
-            entity.setModifyUserName(user.getName());
-        }
-        if (requestArgs != null) {
+        entity.setModifyUserID(user.getId());
+        entity.setModifyUserName(user.getName());
+        if (requestArgs != null&&requestArgs.length()>0) {
             mRequestArgService.deleteByInterfaceId(interfaceId);
             JSONArray array = new JSONArray(requestArgs);
             for (int i = 0; i < array.length(); i++) {
@@ -212,12 +227,12 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
                 }
             }
         }
-        if (responseArgs != null) {
+        if (responseArgs != null&&responseArgs.length()>0) {
             mResponseArgService.deleteByInterfaceId(interfaceId);
             JSONArray array = new JSONArray(responseArgs);
             for (int i = 0; i < array.length(); i++) {
                 JSONObject jsonObject = array.getJSONObject(i);
-                saveResponseArg(jsonObject, projectId, user.getId(), user.getName(), "0");
+                saveResponseArg(jsonObject, projectId, interfaceId, user.getId(), user.getName(), "0");
             }
         }
         try {
@@ -267,6 +282,107 @@ public class InterfaceServiceImpl extends BaseServiceImpl<InterfaceEntity> imple
         map.put("projectName", entity.getProjectName());
         map.put("createTime", DataUtils.formatDate(entity.getCreateTime()));
         map.put("createUserName", entity.getCreateUserName());
+
+        List<RequestArgEntity> requestArgEntities = mRequestArgService.executeCriteria(ResponseArgUtils.getArgByInterfaceId(interfaceId));
+        if (requestArgEntities != null) {
+            List<Map<String,Object>> request = new ArrayList<Map<String, Object>>();
+            for (RequestArgEntity requestArgEntity : requestArgEntities) {
+                MapUtils m = new MapUtils();
+                m.put("id", requestArgEntity.getId());
+                m.put("name", requestArgEntity.getName());
+                m.put("pid", requestArgEntity.getPid());
+                m.put("typeId", requestArgEntity.getTypeId());
+                m.put("note", requestArgEntity.getNote());
+                request.add(m.build());
+            }
+            map.put("requestArgs", request);
+        }
+        List<ResponseArgEntity> responseArgEntities = mResponseArgService.executeCriteria(ResponseArgUtils.getArgByInterfaceId(interfaceId));
+        if (responseArgEntities != null) {
+            List<Map<String,Object>> response = new ArrayList<Map<String, Object>>();
+            for (ResponseArgEntity requestArgEntity : responseArgEntities) {
+                MapUtils m = new MapUtils();
+                m.put("id", requestArgEntity.getId());
+                m.put("name", requestArgEntity.getName());
+                m.put("pid", requestArgEntity.getPid());
+                m.put("typeId", requestArgEntity.getTypeId());
+                m.put("note", requestArgEntity.getNote());
+                response.add(m.build());
+            }
+            map.put("responseArgs", response);
+        }
         return new BaseResult(1, "ok", map.build());
     }
+
+    @Override
+    public ResponseEntity<byte[]> download(String projectId, String userId) {
+        UserEntity user = mUserService.get(userId);
+        if (user == null) {
+            return null;
+        }
+
+        ProjectEntity project  = mProjectService.get(projectId);
+        if (project == null) {
+            return null;
+        }
+
+        JSONStringer stringer = new JSONStringer();
+        stringer.object().key("projectName").value(project.getName())
+                .key("note").value(project.getNote())
+                .key("property").value(project.getProperty())
+                .key("createTime").value(DataUtils.formatDate(project.getCreateTime()));
+        List<InterfaceEntity> list = getBaseDao().executeCriteria(InterfaceUtils.getInterfaceByProjectId(projectId));
+        if (list == null) {
+            return null;
+        }
+/*        for (InterfaceEntity entity : list) {
+            map.put("id", entity.getId());
+            map.put("name", entity.getName());
+            map.put("method", entity.getHttpMethodName());
+            map.put("group", entity.getGroupName());
+
+            List<RequestArgEntity> requestArgEntities = mRequestArgService.executeCriteria(ResponseArgUtils.getArgByInterfaceId(entity.getId()));
+            if (requestArgEntities != null) {
+                List<Map<String,Object>> request = new ArrayList<Map<String, Object>>();
+                for (RequestArgEntity requestArgEntity : requestArgEntities) {
+                    MapUtils m = new MapUtils();
+                    m.put("id", requestArgEntity.getId());
+                    m.put("name", requestArgEntity.getName());
+                    m.put("pid", requestArgEntity.getPid());
+                    m.put("typeId", requestArgEntity.getTypeId());
+                    m.put("note", requestArgEntity.getNote());
+                    request.add(m.build());
+                }
+                map.put("requestArgs", request);
+            }
+            List<ResponseArgEntity> responseArgEntities = mResponseArgService.executeCriteria(ResponseArgUtils.getArgByInterfaceId(entity.getId()));
+            if (responseArgEntities != null) {
+                List<Map<String,Object>> response = new ArrayList<Map<String, Object>>();
+                for (ResponseArgEntity requestArgEntity : responseArgEntities) {
+                    MapUtils m = new MapUtils();
+                    m.put("id", requestArgEntity.getId());
+                    m.put("name", requestArgEntity.getName());
+                    m.put("pid", requestArgEntity.getPid());
+                    m.put("typeId", requestArgEntity.getTypeId());
+                    m.put("note", requestArgEntity.getNote());
+                    response.add(m.build());
+                }
+                map.put("responseArgs", response);
+            }
+        }*/
+        stringer.endObject();
+        try {
+            String path = "I://";
+            String fileName = me.zhouzhuo810.zzapidoc.common.utils.FileUtils.saveFileToServer(stringer.toString(), path);
+            HttpHeaders headers = new HttpHeaders();
+            headers.setContentType(MediaType.APPLICATION_OCTET_STREAM);
+            headers.setContentDispositionFormData("attachment", fileName);
+            return new ResponseEntity<byte[]>(FileUtils.readFileToByteArray(new File(path+fileName)), headers, HttpStatus.CREATED);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return null;
+
+    }
+
 }
